Fix remaining clippy warnings in arkavo-cli
- Add conditional tracing imports for macOS mcp-tools feature
- Remove duplicate std::fmt::Write import
- Convert tools list building to use fold with write!
- Allow unnecessary_cast for cross-platform disk space calculation

+++ crates/arkavo-cli/src/first_run.rs
@@ -245,7 +245,8 @@ pub fn get_available_disk_space() -> f64 {
         let mut stat: MaybeUninit<libc::statvfs> = MaybeUninit::uninit();
         if libc::statvfs(path_cstr.as_ptr(), stat.as_mut_ptr()) == 0 {
             let stat = stat.assume_init();
-            let available = stat.f_bavail * stat.f_frsize;
+            #[allow(clippy::unnecessary_cast)] // Types differ between macOS (u32) and Linux (u64)
+            let available = stat.f_bavail as u64 * stat.f_frsize as u64;
             return available as f64 / 1_000_000_000.0;
         }
     }
